Fix for fetching pubkey when building against 8.0
This is tesed with:
- Compiled and ran with 8.0.18 libraries
- Server running with 8.0.17
- Account using `caching_sha2` auth

The problem is this: We check if `MYSQL_OPT_GET_SERVER_PUBLIC_KEY` is defined.
This is expected to succeed on 8.0.18, but it is not.
Now replacing it with a version check.

To test:
```
use v5.30.0;
use DBI;
my $dbh = DBI->connect("DBI:mysql:host=127.0.0.1;port=8017;database=test;mysql_ssl=0;mysql_get_server_pubkey=1",
                       "msandbox", "msandbox", {'RaiseError' => 1});
my $sth = $dbh->prepare('SHOW SESSION STATUS LIKE "Ssl_cipher"');
$sth->execute();
while (my $row = $sth->fetchrow_arrayref()) {
  say $row->[0] . " = " . $row->[1];
}
$sth->finish();
$dbh->disconnect();
```

Use `RESTART` to restart the server to ensure you are not using the caching features of `caching_sha2`.

* With `mysql_ssl=0;mysql_get_server_pubkey=0` the connection should fail.
* With `mysql_ssl=0;mysql_get_server_pubkey=1` the connection should succeed, with no SSL cipher shown.
* With `mysql_ssl=1;mysql_get_server_pubkey=0` the connection should succeed, with a SSL cipher shown.
* With `mysql_ssl=1;mysql_get_server_pubkey=1` the connection should succeed, with a SSL cipher shown.